Add about menu item and make sure settings retatained after each game session

+++ Task 3 Template/src/de/unikl/seda/snake/gui/snake/menu/interfaces/Adjustable.java
@@ -9,7 +9,7 @@ public abstract class Adjustable implements MenuItem {
         this.snakeGameSettingsAdjuster = snakeGameSettingsAdjuster;
     }
 
-    public abstract int getValue();
+    public abstract double getValue();
     public abstract void increase();
     public abstract void decrease();
 }
